Refactor dbStatements to the db package
