+++ core/src/jvmMain/kotlin/de/p7s1/qa/sevenfacette/db/DatabaseFactory.kt
@@ -1,6 +1,5 @@
 package de.p7s1.qa.sevenfacette.db
 
-import de.p7s1.qa.sevenfacette.DbStatements
 import org.jetbrains.exposed.sql.Database
 import org.jetbrains.exposed.sql.transactions.TransactionManager
 import org.jetbrains.exposed.sql.transactions.transaction
