+++ example-js/src/main/kotlin/de.p7s1.qa.sevenfacette/JSApp.kt
@@ -4,7 +4,6 @@ import de.p7s1.qa.sevenfacette.veritas.all
 import de.p7s1.qa.sevenfacette.veritas.verifyThat
 import de.p7s1.qa.sevenfacette.veritas.verification.hasLength
 import de.p7s1.qa.sevenfacette.veritas.verification.startsWith
-import mu.KLogger
 import mu.KotlinLogging
 
 
@@ -13,7 +12,7 @@ fun main() {
 
     logger.info { "Hello World!" }
 
-    val statementForExecution = DbStatements()
+    val statementForExecution = de.p7s1.qa.sevenfacette.db.DbStatements()
     statementForExecution.add("statement")
     println(statementForExecution.statements)
