+++ example-jvm/src/main/java/de/p7s1/qa/sevenfacette/ExampleApplication.kt
@@ -31,12 +31,8 @@ class ServerApplication
 fun main(args: Array<String>) {
     val logger = LoggerFactory.getLogger(ServerApplication::class.java)
 
-    //val reader:ScriptReader = ScriptReader("/database/cc/test.sql")
-    //val content = reader.getStatements()
+    // Reading SQL-Scripts from Resources
     val content = ScriptReader().getStatements("/database/cc/test.sql")
-
-
-
 
     // Start Database handling
